Added some exception handling to the CLI connector

+++ totalopenstation-cli-connector.py
@@ -1,10 +1,11 @@
 #! /usr/bin/env python
 # -*- coding: utf-8 -*-
-# filename: interactive.py
+# filename: totalopenstation-cli-connector.py
 # Copyright 2008 Stefano Costa <steko@iosa.it>
 # Under the GNU GPL 3 License
 
 import sys
+import os
 
 from optparse import OptionParser
 
@@ -38,7 +39,10 @@ parser.add_option("-o",
 
 (options, args) = parser.parse_args()
 
-exec('from models.%s import ModelConnector' % options.model)
+try:
+    exec('from models.%s import ModelConnector' % options.model)
+except ImportError, message:
+    sys.exit("\nError:\n%s\n" % message)
 
 station = ModelConnector(options.port)
 station.open()
@@ -48,10 +52,13 @@ print "Start download from %s device" % options.model
 result = station.fast_download()
 
 if options.outfile:
-    e = open(options.outfile, 'w')
-    e.write(result)
-    e.close()
-    print "Downloaded data saved to out file %s" % options.outfile
+    if not os.path.exists(options.outfile):
+        e = open(options.outfile, 'w')
+        e.write(result)
+        e.close()
+        print "Downloaded data saved to out file %s" % options.outfile
+    else:
+        sys.exit("Specified output file already exists\n")
 else:
     sys.stdout.write(result)
